requirements.txt
langgraph>=0.2.0
langchain>=0.3.0 
ipython>=8.0.0 
langchain_huggingface>=0.3.0 
langchain_community>=0.3.0 
python-dotenv>=1.0.0 
chromadb>=0.5.0 
langchain_chroma>=0.1.0 
sentence-transformers>=3.0.0 
transformers>=4.40.0